requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.1.0
psycopg==3.2.3
sqlparse==0.5.2
typing_extensions==4.12.2
